fix: remove torch from dependencies in pyproject.toml
