+++ src/modules/speech/tts/pyttsx3_tts.py
@@ -5,7 +5,7 @@ import os
 from pydub.utils import mediainfo
 from pydub import AudioSegment
 
-from src.common.utils.audio_utils import read_audio_file
+from src.common.utils.wav import read_audio_file
 from src.common.interface import ITts
 from src.common.session import Session
 from src.common.types import Pyttsx3TTSArgs, PYTTSX3_SYNTHESIS_FILE, RECORDS_DIR
