+++ src/modules/speech/vad_analyzer/base.py
@@ -1,4 +1,4 @@
-from src.common.utils.audio_utils import calculate_audio_volume, exp_smoothing
+from src.common.utils.helper import exp_smoothing, calculate_audio_volume
 from src.common.types import VADAnalyzerArgs, VADState
 from src.common.interface import IVADAnalyzer
 from src.common.factory import EngineClass
