+++ test/modules/speech/asr/test_whisper_asr.py
@@ -25,7 +25,8 @@ import asyncio
 
 
 from src.common.logger import Logger
-from src.common.utils.audio_utils import save_audio_to_file, load_json, get_audio_segment
+from src.common.utils.helper import load_json, get_audio_segment
+from src.common.utils.wav import save_audio_to_file
 from src.common.factory import EngineFactory, EngineClass
 from src.common.session import Session
 from src.common.interface import IAsr
