+++ test/modules/speech/detector/test_pyannote.py
@@ -7,7 +7,8 @@ import logging
 import unittest
 
 from src.common.logger import Logger
-from src.common.utils.audio_utils import load_json, get_audio_segment, bytes2TorchTensorWith16
+from src.common.utils.helper import load_json, get_audio_segment
+from src.common.utils.audio_utils import bytes2TorchTensorWith16
 from src.common.factory import EngineFactory
 from src.common.session import Session
 from src.common.types import SessionCtx, TEST_DIR, MODELS_DIR, RECORDS_DIR
