+++ test/modules/speech/detector/test_webrtc_silero_vad.py
@@ -4,7 +4,8 @@ import logging
 
 import unittest
 
-from src.common.utils.audio_utils import get_audio_segment, save_audio_to_file
+from src.common.utils.helper import get_audio_segment
+from src.common.utils.wav import save_audio_to_file
 from src.common.logger import Logger
 from src.common.factory import EngineFactory, EngineClass
 from src.common.session import Session
