+++ test/modules/speech/detector/test_webrtc_vad.py
@@ -4,7 +4,9 @@ import logging
 
 import unittest
 
-from src.common.utils.audio_utils import get_audio_segment, save_audio_to_file, convert_sampling_rate_to_16k, read_audio_file
+from src.common.utils.audio_utils import convert_sampling_rate_to_16k
+from src.common.utils.helper import get_audio_segment
+from src.common.utils.wav import save_audio_to_file, read_audio_file
 from src.common.logger import Logger
 from src.common.factory import EngineFactory, EngineClass
 from src.common.session import Session
