+++ test/modules/speech/recorder/test_vad_record.py
@@ -11,7 +11,7 @@ from src.common.interface import IAudioStream, IRecorder
 from src.common.logger import Logger
 from src.common.factory import EngineFactory, EngineClass
 from src.common.session import Session
-from src.common.utils import audio_utils
+from src.common.utils import wav
 from src.common.types import SessionCtx, MODELS_DIR, RECORDS_DIR
 
 import src.modules.speech.detector.porcupine
@@ -90,7 +90,7 @@ class TestVADRecorder(unittest.TestCase):
         frames = asyncio.run(self.recorder.record_audio(self.session))
         self.assertGreater(len(frames), 0)
         data = b''.join(frames)
-        file_path = asyncio.run(audio_utils.save_audio_to_file(
+        file_path = asyncio.run(wav.save_audio_to_file(
             data, os.path.join(RECORDS_DIR, "test.wav")))
         print(file_path)
 
@@ -98,7 +98,7 @@ class TestVADRecorder(unittest.TestCase):
         frames = asyncio.run(self.recorder.record_audio(self.session))
         self.assertGreater(len(frames), 0)
         data = b''.join(frames)
-        file_path = asyncio.run(audio_utils.save_audio_to_file(
+        file_path = asyncio.run(wav.save_audio_to_file(
             data, os.path.join(RECORDS_DIR, "test.wav")))
         print(file_path)
 
@@ -108,7 +108,7 @@ class TestVADRecorder(unittest.TestCase):
         frames = asyncio.run(self.recorder2.record_audio(self.session))
         self.assertGreater(len(frames), 0)
         data = b''.join(frames)
-        file_path = asyncio.run(audio_utils.save_audio_to_file(
+        file_path = asyncio.run(wav.save_audio_to_file(
             data, os.path.join(RECORDS_DIR, "test2.wav")))
         print(file_path)
         self.recorder2.close()
@@ -132,6 +132,6 @@ class TestVADRecorder(unittest.TestCase):
             frames = asyncio.run(self.recorder.record_audio(self.session))
             self.assertGreaterEqual(len(frames), 0)
             data = b''.join(frames)
-            file_path = asyncio.run(audio_utils.save_audio_to_file(
+            file_path = asyncio.run(wav.save_audio_to_file(
                 data, os.path.join(RECORDS_DIR, f"test{i}.wav")))
             print(file_path)
